+++ tools/themes/tool.js
@@ -8,7 +8,7 @@
         name: 'themes',
         category: 'general',
         type: 'multiple',
-        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'pieceDrag','noPractice'/*, 'noGrab'*/],
+        possibleValues: ['performance', 'justExplorer', 'mobile', 'slimArrows', 'flairX', 'lessIcons', 'nonStickyHeader', 'noStudyChat', 'pieceDrag','noPractice', 'gameMoveList'/*, 'noGrab'*/],
         defaultValue: '',
         advanced: true
       }
@@ -29,7 +29,8 @@
         'themes.noStudyChat': 'No study chat',
         'themes.pieceDrag': 'Nicer piece drag',
         'themes.noGrab': 'No grab cursor',
-        'themes.noPractice': 'No Practice button'
+        'themes.noPractice': 'No Practice button',
+        'themes.gameMoveList': 'Flexible game move list'
       },
       'ro-RO': {
         'options.general': 'General',
@@ -45,7 +46,8 @@
         'themes.noStudyChat': 'F\u0103r\u0103 chat \u00een studii',
         'themes.pieceDrag': 'Apucat piese mai frumos',
         'themes.noGrab': 'F\u0103r\u0103 cursor care apuc\u0103',
-        'themes.noPractice': 'F\u0103r\u0103 buton Antrenament cu calculatorul'
+        'themes.noPractice': 'F\u0103r\u0103 buton Antrenament cu calculatorul',
+        'themes.gameMoveList': 'List\u0103 mut\u0103ri flexibil\u0103 \u00een joc'
       }
     }